services/export_service/requirements.txt
aiohttp
aioboto3
pydantic
pyyaml
aiogremlin
pydantic-settings
